# Optimize gaussian op for scalar matrix