As suggested by @fritzo , we can improve the speed of GaussianHMM for the case `trans_dim, obs_dim == (1, 1)` by using scalar ops instead of linear algebra ops. This PR incorporates that enhancement.

Current tests (with trans_dim=obs_dim=1) already cover the implementation.

Profiling result: with
```
num_steps, obs_dim, hidden_dim = 6000, 1, 1
sample_shape = (5,)
batch_shape = (4,)
```
+ forward+backward: 52.9 ms (vs non-optimize one: 191 ms)
+ only forward: 12 ms (vs non-optimize one: 52.7 ms)

So we can observe 4x improvement with this PR.